smartlearn/ai-worker/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.1
redis==5.0.1
requests==2.31.0
pymupdf==1.23.8
pdfplumber==0.10.3
python-dotenv==1.0.0
jsonschema==4.20.0
numpy==1.26.3
groq==0.9.0
google-generativeai==0.3.2
python-multipart==0.0.6
langchain-core==0.1.10
langchain-community==0.0.13
aiofiles==23.2.1
httpx==0.25.2
openai==1.12.0
